lace_stats tests passing
